amplifierd/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amplifierd"
version = "0.1.0"
description = "Amplifier daemon - HTTP API for the Amplifier AI platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "amplifier-library",
    "amplifier-foundation",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "sse-starlette>=2.2.1",
    "amplifier-core",
    "gitpython>=3.1.45",
    "fsspec>=2025.10.0",
    "claude-code-sdk>=0.0.25",
    "requests>=2.32.4",
    "aiohttp>=3.13.2",
    "anthropic>=0.75.0",
    "bs4>=0.0.2",
    "apscheduler>=3.10.0",
    "psutil>=5.9.0",
    "click>=8.1.0",
]

[project.scripts]
lakehouse = "amplifierd.cli:main"
amplifierd = "amplifierd.__main__:main"

[tool.setuptools.packages.find]
include = ["amplifierd*"]

[tool.uv.workspace]
# Add all projects in the workspace
members = []

[tool.uv.sources]
amplifier-core = { git = "https://github.com/microsoft/amplifier-core", branch = "main" }
amplifier-library = { path = "../amplifier_library", editable = true }
amplifier-foundation = { path = "../related-projects/amplifier-foundation", editable = true }

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "debugpy>=1.8.14",
    "httpx>=0.28.1",
    "pyright>=1.1.407",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
    "twine>=6.1.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [
    "**/__pycache__",
    "**/.venv/**",
    "**/node_modules/**",
    ".git/**",
    ".ruff_cache/**",
    ".pytest_cache/**",
    "*.egg-info",
    "build/**",
    "dist/**",
    ".data/**",
    "ai_working/**",
    ".amplifierd/**",
]
typeCheckingMode = "basic"
reportMissingImports = false
pythonVersion = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests for library layer",
    "integration: Integration tests for daemon layer",
]
